Tests/LibWeb/Text/expected/selection-drag-above-viewport.txt

selection reached viewport top: true
above viewport matches viewport top: true
above viewport selected forward text: false
